+++ src/django_twc_toolbox/crud/views.py
@@ -95,35 +95,14 @@ class CRUDView(NeapolitanCRUDView):
         if filterset is not None:
             queryset = filterset.qs
 
-        # set object_list up here instead of down in the `if paginate_by is None` checks
-        self.object_list = queryset
-
         if not self.allow_empty and not queryset.exists():
             raise Http404
 
-        if self.table_class is not None:
-            # If `table_class` is set, that means the view class has `SingleTableMixin` in it's inheritance
-            # chain[^1]. `neapolitan.views.CRUDView.get_paginate_by` does not take an argument, whereas
-            # `django_tables2.views.SingleTableMixin.get_paginate_by` expects one (`table_data`), so we need
-            # to pass in an argument so the view doesn't crash.
-            #
-            # Now, mind you, `SingleTableMixin.get_paginate_by` doesn't appear actually **do** anything with
-            # that argument, so it seems like a bit of a useless thing. But 🤷 whatareyagonnado? I'm sure they
-            # have their reasons when they wrote it that way.
-            #
-            # I have opened an issue on neapolitan to add the ability for `CRUDView.get_paginate_by` to take
-            # arbitrary args and kwargs by adding `*args, **kwargs`. I may also open an issue on django-tables2
-            # to add `None` to the `table_data` argument in `SingleTableMixin.get_paginate_by` since it's unused
-            # and presumably only put there to allow people to override and customize?
-            #
-            # [^1]: Inheritance chain? Is that the correct terminology? Probably not, but you get the idea.
+        self.object_list = queryset
 
-            paginate_by = self.get_paginate_by(self.object_list)  # type: ignore[call-arg]  # pyright: ignore[reportCallIssue,reportUnknownVariableType]
-        else:
-            paginate_by = self.get_paginate_by()
+        paginate_by = self.get_paginate_by(self.object_list)
 
         if paginate_by is None:
-            # Unpaginated response
             context = self.get_context_data(
                 page_obj=None,
                 is_paginated=False,
@@ -131,7 +110,6 @@ class CRUDView(NeapolitanCRUDView):
                 filterset=filterset,
             )
         else:
-            # Paginated response
             page = self.paginate_queryset(self.object_list, paginate_by)
             context = self.get_context_data(
                 page_obj=page,
@@ -141,6 +119,10 @@ class CRUDView(NeapolitanCRUDView):
             )
 
         return self.render_to_response(context)
+
+    @override
+    def get_paginate_by(self, *args: object, **kwargs: object) -> int | None:
+        return super().get_paginate_by()
 
     @override
     def get_context_data(self, **kwargs: object) -> dict[str, object]:
@@ -155,6 +137,7 @@ class CRUDView(NeapolitanCRUDView):
     @override
     def get_template_names(self):
         template_names = super().get_template_names()
+
         # only render the template partial if:
         # - it's the list view
         # - it's an HTMX request
@@ -174,6 +157,7 @@ class CRUDView(NeapolitanCRUDView):
                 f"{template_name}#{self.list_partial}"
                 for template_name in template_names
             ]
+
         return template_names
 
     @classmethod